move cipher-agnostic code to cipher.go
